+++ packages/webos/CHANGELOG.md
@@ -13,6 +13,11 @@ No significant changes.
 ## [4.7.11] - 2024-05-13
 
 No significant changes.
+## [unreleased]
+
+### Removed
+
+- `webos/platform` member `watch` and `legacy`
 
 ## [4.9.0-alpha.1] - 2024-04-09